feat: add tabGroupIndicator theme property and extract wrap underline logic
Add a `tabGroupIndicator` property to `DockviewTheme` that controls how
tab groups are visually indicated: 'wrap' (default SVG wrap-around) or
'none' (pure CSS via tab classes and --dv-tab-group-color custom property).

Extract the ~480-line SVG wrap underline logic from TabGroupManager into
a self-contained WrapTabGroupIndicator class behind an ITabGroupIndicator
interface, reducing TabGroupManager from 892 to 541 lines.

Set the abyss theme to use 'none' mode with a pure CSS box-shadow
indicator as a demonstration of CSS-only tab group styling.
